requirements.txt
flask==3.1.1
flask-cors==5.0.1
openai==1.78.1
python-dotenv==1.1.0
gunicorn==21.2.0 